requirements.txt
telethon 
asyncpraw
request
bs4
python-dotenv